requirements.txt
os
numpy
collections
matplotlib.pyplot
wordcloud
csv
pandas
